requirements.txt
osc-sdk-python==0.9.13
ansible==2.8.5
ansible-subprocess==0.4.0
Flask==1.1.1
Flask-Caching==1.7.2
pyzmq==18.1.0
SQLAlchemy==1.3.8